chore: cleaned up types and services, seperated types and services from machine files
